Let blocked threads have an actual fair chance to be scheduled
Just yielding the thread is insufficient to allow blocked threads to be
scheduled. Since we process buckets in chunks of 10k for "background"
iterations, a microsecond sleep should have a minimal impact in
practice.

A more thorough fix would involve redesigning how locking and signalling
is done in the bucket DB. This simple fix should however address the
blocking issues observed with the current design.