VIP status for combined clusters
When a container and content cluster is set up such that
nodes are paired 1-1 and all content nodes have the entire corpus,
containers can bypass dispatching and always search the local
peer content node ("direct search").

In this case, if the local content node is down or is missing data,
we prefer not receiving traffic on the local container node to avoid
the unnecessary latency and network load from having to route queries
from the local container to a remote content node.

This commit determined VIP status of containers based on the status
of the local content node, instead of the default method, when we
recognize that we have a "combined cluster" configuration.