Reduce log level for "unexpected" wrong distribution reply to debug
Client load can be rejected towards distributors even with a matching
cluster state version. This usually happens during a node fail-over
transition, where the target distributor will reject an operation
bound to a particular bucket if it does not own the bucket in _both_
the current and the next (transition target) state. Since it can happen
during normal operation and will happen per client operation, we keep
this as debug level to prevent spamming the logs.