Implement distributor bucket ownership handover safe time point (#1098)
This adds a period of time after each distributor bucket ownership handover edge where mutations requiring timestamp assignments are bounced back to the client. This is in order to avoid the possibility of generating the same timestamp as a previous distributor whose bucket subset we now own. The time period is configurable in whole seconds based on expected worst-case cluster clock skew and the feature may be disabled entirely by setting the configured value to zero.

Read-only operations and mutating operations that don't require new timestamps (such as RemoveLocation are allowed through).