Don't restrict to only config servers
We want zk access limitation to be deactivated by default for self-hosted applications.
However, these also run the node repo and will would therefore restrict to config servers
only as the node repo is empty. This changes to passing an empty access restriction list
instead if the node repo is empty.