Use latest candidate cluster state when comparing against reported node states
Using just the versioned cluster state instead can cause the code to
erroneously believe that it is seeing repeated reported state changes
for the first time. This happens when the diffs in the reported node
states are not in and by themselves enough to trigger a new cluster state
version containing the changes.

This can in turn spam the logs and event buffers until a new cluster state
has been versioned.