Increase number of buckets/sec processed by distributors
This greatly reduces time to process large bucket databases, especially
when the number of buckets go into the millions. With the previous
defaults, this could take absolute ages and fail to get in sync within a
reasonable time.

If we do not observe any negative effects as part of this change, we'll
likely boost the ticks_before_wait default even more.