Update with new overrides.
Removing old staging override, not used, was intended for old screwdriver config.
Adding two new overrides. One is for enabling the tests while it is default off (in the beginnging)
and one for disabling it after it has been enabled as default.
The enable override should be deprecated after it has been turned on as default.