Add operation sequencing for put, remove and update operations
Keeps track of GIDs of pending operations and bounces operations that
arrive to these GIDs before the original operation has completed.

RemoveLocation operations are currently not handled due to these covering
an a priori unknown set of GIDs, but sequencer can be extended to support
these as well with some extra work.

Enabled by default, but may be disabled via config.