Multinode orchestration
The goal is to allow suspension of several nodes within an application at the
same time.  The same per-cluster constraint policy is enforced, e.g. only some
max percentage of service instances are allowed to suspend at any given time.

In particular, this should solve VESPA-7162, which saw an application having 3
nodes in the same cluster (TLD) on the same Docker host.  The 3 nodes are
packaged into a NodeGroup tied to the application, and if all TLD nodes on
other Docker hosts are up, we're allowed to suspend these 3.

To make it easier to test, I have refactored the model-dependent calculations
into ApplicationApi and ClusterApi: These will do the actual traversing of the
service monitor's datastructure (model), as well as the per-node ZooKeeper flag
(NO_REMARKS/ALLOWED_TO_BE_DOWN).  Together these expose the minimal API for the
policy to base their decision on.

Testing of the policy is currently verbose and hard due to the hard-wiring of
the use of the model.  This PR should make it easy to test the policy, and
separately make the testing of the model-dependent *Api classes as easy as it
gets.

This is a rough first-draft with e.g. TODOs.  Some issues to settle is whether
to make NodeGroup independent of the application instance, the service monitor
model is still used some places, and refactor/write the tests, and more.