Adds classes to give the Orchestrator policy classes a simplified view of Vespa.
This should be a no-op.  The only changes that actually could have an impact
are the changes to getting the cluster controllers, but it should be
functionally equivalent.

This PR will make it easier to change the Orchestrator policy to allow
suspending several nodes (NodeGroup) in an application on a single Docker host.