Make struct builders public.
Before this change, a new struct builder had to be created to set values
on the builder. Then this builder was set on the toplevel config builder,
replacing the previous struct builder and any changes already done.
Now however, it is possible to set only a subset of values on a default constructed
struct builder without affecting any previous changes.