Deprecate and ignore reply-after attribute
"n-of-m" redundancy has too many unfortunate edge cases for consistency
and distributor memory usage to keep this feature around.
Field still remains in config, but is always set equal to final redundancy.

If reply-after attribute is present, a deployment deprecation warning is issued.

If we move to consensus-based consistency, bringing back this feature will
effectively be a no-op.