Orchestrator support for setting host status
Directly setting the host status can be useful for an operator, e.g. to break a
deadlock.

Unfortunately, some code use the term "host status" while others use "host
state".  I haven't settled yet on which is preferred, e.g. 'state' is used in
REST APIs, but 'status' is the original term and slightly more used.  This PR
maintains the local use of the terms, meaning it adds 'state' where that's
normally used and 'status' other places.

Setting the deadlock is done with a PATCH request, which is defined in
jaxrs_utils. jaxrs_utils only defines PATCH, and that's not used anywhere else
except in jaxrs_client_utils tests.  So I remove this module as a dependency a
couple of places.