Inhibit scheduling of merges towards nodes that are marked busy
Utilizes existing maintenance operation scheduler system which checks
if an operation is considered blocked and does not start it if this
is the case. We now block a merge operation if any of the nodes in its
node set are marked busy by the pending message tracker.

Duration for which nodes are marked busy is live-configurable.