Enforce order of block-change
Enforce that the block-change tag is placed after any test and staging tag and before the prod tag.
This is in case we want to allow multiple block points in the future (perhaps to block rolling out
to the next instance when having multiple instances) such that the placement
becomes significant.