Verify the actual change we are deploying
This code verified the change we believe we are currently deploying
according to application.deploying().
Since we now allow new changes to enter while an application change is
already in progress, application.deploying() is not reliable.

Moving the check to after we update application.deploying()
to what we received in this deploy request should solve this
problem.