Also set the distributor wanted state when safe-setting the storage node state
This is done as part of the SAFE REST API call to set the node state of a
storage node to ensure atomicity of the state change, reduce the number of
state changes, and minimize the time to complete the state changes.

The right way to think about the safe-set is then: In order to safely set a
storage node to (e.g.) maintenance, the distributor will also have to be set to
down. And so on for the various permutations of state transitions.