Make SlobrokMonitorManager a component
To get insights into the Slobrok data used by the service monitor, it would be
useful to expose SlobrokMonitorManager. For instance it could provide a list of
name/spec pairs of services that are up (added as ServiceMonitorManager::lookup
in this PR).

The first step towards this is making SlobrokMonitorManager a component.