Remove memory manager component from content layer
We already have resource utilization tracking in both MessageBus
and the search core. The memory manager has never been auto-scaled
based on the hardware present and adds a _lot_ of complexity without
having any known instances where it has actually saved the day.

Removing it also removes a mutex on the message hot path.

If we need such functionality in the future, should design a lock-free
solution.

Cleanup