Always retry hanging jobs
Hanging jobs will usually be implicitly retried, since
they will not be taken as running.

However, they will not be retried if the outcome of completing
the job will be unchanged. Such jobs still need to complete, to trigger
the next step, such that we can eventually trigger the downstream
jobs which need to run to complete the ongoing change
(arguably we could skip the job in these cases, but currently we don't).