Parse generated tensor function trees
To make generated tensor function trees transparent to
the config model we need to convert each tensor function node
to the corresponding ranking expression node.
This is most easily done by parsing the tensor function
tree string output as a ranking expression (something
which is required to always work in any case).