Redefine task and its context
After discussions with bratseth, we'll not organize tasks in a tree structure,
and instead have tasks mostly 1:1 with components. Tasks are put in chains, and
may have ordering constraints between them that the chain resolves.

The FileSystem API has been removed and instead use the raw java NIO
FileSystem, and use Jimfs to mock FileSystem in tests.

NodeAdminStateUpdater has been made an interface, since the HTTP request
handler needs to delegate to it when getting requests. Much of the public HTTP
API will get internalized with standalone NodeAdmin.