Avoid Slobrok in node-admin cluster
When this rolls out, two Slobroks on the node-admin will be removed, and 1
additional Slobrok will be put on one of the proxy nodes if the routing cluster
has 3 or more nodes.

The Orchestrator sees the latest model, i.e. expect 3 Slobroks on the proxy
nodes. Since the 1 additional Slobrok service is down at the start of the
rollout, only that host will be allowed to suspend among the 3. But if that 1
additional Slobrok service comes up quickly, then any of the 3 can suspend. I
think this will just work - no deadlocks or similar.

I couldn't find any test that would covers this code change, but will write
tests if someone can help me with this, if deemed necessary.