Fix certificate attribute name
The filter did not find the certificate in certain cases (possibly JAX-RS APIs),
even though the acess log showed that the client certificate was used.

I assume the reason is that a different attribute name was used. I see
`javax.servlet.request.X509Certificate` is used in other places as well.