Prefix constants by model name
Large constants are cross rank profiles. This avoids name
conflicts when multiple models are used. It is not strictly
necessary because the user can always disambiguate when chosing
names, but there is a scenario where conflicts are plausible and
leaving this to users is inconvenient: Multiple versions of the
"same" model are tested in different rank profiles.