Explicitly check for empty steps
The previous check didn't work with serialization as a deserialized empty
`DeploymentSpec` != `DeploymentSpec.empty` because `DeploymentSpec` does not
implement `equals`.

The tests currently pass because `MemoryControllerDb` just stores references
which satifies `Object.equals`.

Issues like this will be caught when we get rid of `ControllerDb` (which is how
I discovered this).