Handle explicit node state response case without pending request
Deals with the edge case where a node explicitly triggers a node state
response to be sent (generally due to updated host info to the controller(s))
but where there is no pending request to immediately reply to. We now maintain
state per controller which lets us infer whether it needs to immediately
observe a reply when the next request is received for it. This avoids having to
wait for an entire RPC timeout cycle before a response is sent to the controller.

This feature is predominantly to avoid keeping nodes in the default bucket space
in implicit maintenance mode for longer than is necessary when merges in the
global space are completed.