Support cross-document type visiting via /document/v1/ root
Requires `cluster` to be set since we don't have a document type to
auto-infer the target cluster from. Can use `bucketSpace` parameter
to explicitly state the target bucket space to visit (if not given,
implicitly visits the 'default' space).

Note: since we are not bound to a single document type, the field set
used is `[all]`, not `doctype:[document]`. This means that this does
_not_ have parity with non-root Document V1 visitor requests, though
it _does_ have parity with legacy `/visit` and `vespa-visit`. To
have same behavior for a single document type, use an explicit document
`selection=mydoctype` parameter combined with a `fieldSet` parameter
of `mydoctype:[document]`.

This fixes #5794