Only count time with no threads available
Found this edge-case with logAndDie in ThreadPoolProvider:
 - A Runnable executes and finishes at time T0
 - No Runnables execute for a long time t > maxThreadExecutionTimeSeconds
 - At time T1 = T0 + t, a bursts of Runnables N > maxthreads arrives such that
   the (maxthreads+1) Runnable will be tried executed and rejected before any
   of the first maxthreads Runnables finish.
 - Then, logOrDie will be called immediately.

That is, a sufficiently silent period followed by a burst of Runnables cause
the process to die.

This PR will make sure that the process dies only if
maxThreadExecutionTimeSeconds seconds pass with no available threads, like the
documentation says.

I found this as I was looking for reasons why the Cluster Controller have had
logAndDie problems. The CC has maxthreads of 10, and was getting bursts of 6-8
of a particular request type (reduced in PR 6207).