Avoid fatal first CC request timeout
If the first setNodeState to the "first" cluster controller times out, then
we'd like to leave enough time to try the second CC. This avoids making a
single CC a single point of failure.

The strategy is to set a timeout of 50% of the remaining time, so if everything
times out the timeouts would roughly be 50%, 25%, and 12.5% of original
timeout.

An alternative strategy would be to use 33% for each, which would be more
democratic.