Disable loading minimal set of config model versions in prod
AccessControlValidator uses the existence of a model for the version
it validates to check if this is a first time deployment. If config
server has loaded the minimal set of versions (1, 3) and there comes a new deployment
with version 2 (e.g. because we have just released 3 and it is not rolled
out completely yet), there will exist no version 2 and AccessControlValidator
will think that this is a firsst time deployment and fail if access control
has not been enabled.