Always use a fallback provisioner in hosted Vespa
If we introduce a new cluster or change the number of nodes in an
application between versions we will try to load the model (at startup,
so ActivatedModelsBuilder will be used) and discover that we need to provision
additional nodes. Since ActivatedModelsBuilder had set the fallback provisioner
to null this failed spectacularly. Make sure to always use node repository
provisioner as fallback in this case.