HW verification of IP
Hosts with HOST_NETWORK fails hardware divergence check of IP addresses because
the reverse lookup returns the node hostname (this is the whole point of the
host network - the container requires reverse lookup to pass, but not the
host). For HOST_NETWORK we'll disable this check.

We may want to add verification of reverse lookup of container IP addresses
later.

To compensate, forward resolution (hostname -> IP addresses) has been added.