Segment 'contains' text by default
The current default is to not segment text given to 'contains'
in YQL outside of userInput, which means that submitting text
that is not pre-tokenized in exactly the same way as the tokenizer
of the field won't work (such as the queries "contains 'foo bar'",
or contains foo.bar).

Because of this default choice, apparently some complicated logic was
added to track when it needs to be automatically replaced with a more
reasonable choice, in the case where one Vespa instance federates to
another. I removed that now, because it doesn't seem necessary without
the bad default.