Retry twice if only 1 CC
Caught in the systemtests: If there's only one CC, there will only be 1 request
with timeout ~5s, whereas today the real timeout is 10s. This appears to make a
difference to the systemtests as converging to a cluster state may take several
seconds. There are 2 solutions:
 1. Allocate ~10s to CC call, or
 2. Make another ~5s call to CC if the first one fails.

(2) is simpler to implement for now. To implement (1), the timeout calculation
could receive the number of backends as a parameter, but that would make the
already complex logic here even worse. Or, we could only reserve enough time
for 1 call (abandon 2 calls logic). TBD later.