Send probe when suspending many nodes
When suspending all nodes on a host, first do a suspend-all probe that will try
to suspend the nodes as normal in Orchestrator and cluster controller, but
actually not commit anything. A probe failure will result in the same failure
as a non-probe failure: A 409 response with description is sent back to the
client.