Add flags module
FileFlagSource reads flags from files in /etc/vespa/flags and is a component
that can be injected in host admin, config server, etc. A flag named foo
corresponds to filename foo.

In general a FlagSource manages:
 - Feature flags: A feature is either set (true/enabled) or otherwise false.
   Touching a file foo means the feature flag foo is set (true).
 - Value flags: Either a String or empty if not set. The String corresponds to
   the file content.

The plan is to make the config server another source of flags. A unified
FlagSource can merge the two sources with some priority and used in e.g. parts
of node-admin. In other parts one would only have access to the file source.

Defines various flag facades:
 - FeatureFlag: Used to test whether a feature has been enabled or not.
 - IntFlag
 - JacksonFlag: Deserializes JSON to Jackson class, or return default if unset.
 - LongFlag
 - OptionalJacksonFlag: Deserializes JSON to Jackson class, or empty if unset.
 - OptionalStringFlag
 - StringFlag

This is part of removing some of the last Chef recipes. Some minor tweaks have
been necessary as part of this and are included in this PR (test whether a
systemd service exists, task-friendly file deletion, allow capitalized letters
in YUM package name).