Support feature flags in node repository
This implements feature flags for the node repository. A feature flag can be
toggled on/off for the following dimensions:

1) The node repository (entire zone)
2) A specific node
3) A specific application

Flags must be declared in the `FlagId` enum, this is typically done when
implementing the feature that should be guarded by a flag.

Flag status is stored in ZooKeeper. Inspecting and toggling flag status is done
through a REST API, see `RestApiTest#test_flags()`.