Avoid periodic removal
The controller log contains errors about refusing to handle the config server
application since the controller application is active.

The message is misleading: Only the controller application will be activated,
while the config server application will be tried removed, and the message
comes as part of this removal.

This PR adds a check that actual removal from duper model and provisioner will
only be done if it is active in the duper model.

There are edge cases that the application may be active in the provisioner
while not in duper model, but it's fixed on a restart.