Ensure that asciistream moves and swaps have expected semantics
Defaulted move ctor and assignment will not have the expected behavior
for the current _rbuf pointer when it points into a short-string
optimized _wbuf buffer. I.e. it will be pointing into the buffer in
the object that was just moved away from.

Update swap() to give the correct semantics when either/both arguments
point to a read-only buffer. Would previously reset _rbuf to _wbuf
unconditionally, effectively forgetting the string that was referenced.