Metadata about /state/v1/health status
The service monitor uses /state/v1/health to monitor config servers and the
host admins (but not yet tenant host admins).

This commit adds some metadata about the status of a service:
 - The time the status was last checked
 - The time the status changed to the current

This can be used to e.g. make more intelligent decisions in the Orchestrator,
e.g. only allowing a service to suspend if it has been DOWN longer than X
seconds (to avoid spurious DOWN to break redundancy and uptime guarantees).