Make ModelContext.Properties more accessible in config model
This commit makes it easier to get the ModelContext.Properties instance in the
config model. Before, one would typically have to change a bunch of code to
wire through a new property (builder methods and fields, DeployProperties
constructor, fields, and getters, mock classes, ...).

It is made accessible through:
 - ModelContext.properties(),
 - DeployState.getProperties(), and
 - ConfigModelContext.properties()

The Properties implementation is in the configserver module, making it easy to
back properties with flags.