Make node builder
While trying to add another field to Node I came upon the multitude of places
the constructor is called. This is simplified by using a builder.

The builder is internal to Node for now. If wanted/needed, it can be made
public (especially tests may be interested - I have touched as few as
possible).