Health rest API
Makes a new REST API /orchestrator/v1/health/<ApplicationId> that shows the
list of services that are monitored for health. This information is currently a
bit difficult to infer from
/orchestrator/v1/instances/<ApplicationInstanceReference> since it is the
combined view of health and Slobrok. There are already APIs for Slobrok.

Example content:

$ curl -s localhost:19071/orchestrator/v1/health/hosted-vespa:zone-config-serve\
rs:default|jq .
  {
    "services": [
      {
        "clusterId": "zone-config-servers",
        "serviceType": "configserver",
        "configId": "zone-config-servers/cfg6",
        "status": {
          "serviceStatus": "UP",
          "lastChecked": 1548939111.708718,
          "since": 1548939051.686223,
          "endpoint": "http://cfg4.prod.cd-us-central-1.vespahosted.ne1.yahoo.com:19071/state/v1/health"
        }
      },
      ...
    ]
  }

This view is slightly different from the application model view, just because
that's exactly how the health monitoring is structured (individual monitors
against endpoints).

The "endpoint" information will also be added to /instances if the status comes
from health and not Slobrok.