Use VIP status file for config server (and controller) in hosted
The code for setting up VIP status is already in place in config model,
but depends on value of hostedVespa() in ModelContext.Properties, so have to
make sure that is set for standalone container as well.
Removing config override for vip status, as setting VIP in or out of rotation when not
using a VIP status file is handled when bootstrapping now.