Force node clones to get a copy of all fields
I observed the node reports were removed after some time.

I have traced this to

    https://github.com/vespa-engine/vespa/blob/master/node-repository/src/main/java/com/yahoo/vespa/hosted/provision/Node.java#L134

When Node wants to create a clone of itself with a modification, e.g. with
wantToRetire, it calls that Builder constructor. The problem is that
this.reports is not copied to the Builder's reports field, and so the new node
has lost the report.

A new field 'modelId' has been added to the node after 'reports', and it too
is not copied to the builder.

This is clearly too fragile.

This PR makes a public Builder() constructor taking in ALL fields of the Node
constructor, and forces all code outside to use that Builder constructor by
making the Node constructor private. This is just a refactoring.

However the Builder(Node) constructor calls the new Builder constructor, thus
forcing forwarding of all fields. This solves the problem.