Override method for setting status handler for a config server cluster
Need to use CloudConfigOptions to get correct values for config servers.
Trying to set ModelContext properties did not work as apparently we populate
a lot of the values there by default (which does not seem right)
and there are users of those elsewhere, so one would need to duplicate that
mess in StandaloneContainerApplication as well to make it work.