Add workarounds for legacy global distribution hash handling
This addresses a regression introduced as part of #8479, which in
turn was intended to serve as a fix for issue #8475. This regression
would stall cluster state convergence when a subset of nodes contained
the fix and another subset did not.

With the workarounds present, nodes gracefully handle the case where
different distribution hashes are expected for the global bucket space.

`BucketManager` will now fall back to comparing the new incoming hash
to that of the legacy derived distribution config if it mismatches.

`PendingClusterState` will try to send a subset of bucket info requests
with legacy hash format for the global bucket space iff there has been
at least 1 rejected request.

All these workarounds will be removed on Vespa 8.