Do RPC call asynchronously and with higher timeout
The RPC call is just a hint that the host wants to download the
file reference, no need to wait for call to finish, since we allow
calls to fail. Increase timeout a bit, since we have seen that 100 ms
might not always be enough (e.g. in system tests)