Minimal template class
The Velocity template engine used in Template.java is a terrible fit for shell
scripts, as their two special characters $ and # are frequent in both. This means shell scripts must escape $ and #, sometimes, and the rules for when and how is extremely difficult. I gave up on ${0##*/}, which apparently requires 2 escapes: one for $ and one for double-#. Escaping the double-# can be done with #[

All I want is to make simple substitutions. "Templar" provides that.