Avoid dereferencing symlink
The call

  install_symlink(conf/configserver-app/components lib/jars/config-models)

would on first invocation install the symlink at lib/jars/config-models to
point at .../components. This is what we want.

But if invoked again, it would note config-models is a directory, and therefore
prepare to install a symlink in that directory. The name of the symlink would
be taken from the target (components). And therefore
conf/configserver-app/components/components would point to
conf/configserver-app/components!

With -n, symlinks targets are not dereferenced and therefore the second
invocation should be a no-op (with -f).