Handle getting new config after getting empty config
Sentinel config is empty with generation 0 if an application does not
exist. When we get that, make sure not to cache the config and make
sure to ask for new new config even when we have one for generation 0,
to handle cases where we get empty sentinel config due to some config server
issue/bug when upgrading and where the effect is that empty sentinel config
is served.

This change should make the sentinel start services immediately after such
an issue has been fixed.