Avoid resurrecting replicas for nodes that are unavailable in pending state
We previously only checked for node availability in the _active_ state
without looking at the pending state. This opened up for a race condition
where a reply for a previously DB-pruned node could bring a replica back
in the DB iff received during a pending state window.

Consider Maintenance as unavailable for this case, not just Down.

Also move all `PutOperation` tests to GTest.