Tenant hosts only have node-admin service
Tenant hosts are allocated to the zone (routing) app in the node-admin
cluster. These hosts also have various other services like logd,
config-sentinel, etc that do not run.

The metricsproxy container is a new service that will be added to all hosts,
and that will be Slobrok monitored. Except, it will not run on tenant hosts as
above. To avoid having metrics cluster spanning the whole of the zone app, and
have them DOWN on all tenant hosts, we'll filter those service away from the
model when generating the ApplicationInstance of the service model.