Fix bug in lid space compaction where partial updates to attributes were lost when moving a document.
Instead of using the document store directly (when reading the document to move) the document retriever must be used.
The document retriever patches in attribute values and also correctly waits for the thread writing to the document store.

The job moving documents between the "ready" and "not ready" sub databases was already doing this correctly.