Remove serialization of ClusterSpec rotations
This data is only stored when `<rotations/>` is specified through
`services.xml`. Since we're removing that element, we can also remove the
serialization. It's not used in our shared routing layer.

The `rotations` field is still present in the `/loadbalancers/v1/` response as
the client of this API still expects it.