Reduce ZK session timeout for master election unit tests
Setup methods would previously override explicit timeouts set by the
tests with a default of 2 minutes. It's suspected that this long timeout
combined with spurious ZK client failures is a source of test failures.

Reduce to 10 seconds for now. Let's see how well it fares on a heavily
loaded CI container.